requirements.txt
autopep8==2.0.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
filelock==3.8.0
huggingface-hub==0.11.0
idna==3.4
joblib==1.2.0
nltk==3.7
numpy==1.23.5
packaging==21.3
pandas==1.5.1
Pillow==9.3.0
pycodestyle==2.9.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.6
PyYAML==6.0
regex==2022.10.31
requests==2.28.1
scikit-learn==1.1.3
scipy==1.9.3
sentence-transformers==2.2.2
sentencepiece==0.1.97
six==1.16.0
threadpoolctl==3.1.0
tokenizers==0.13.2
tomli==2.0.1
torch==1.13.0
torchvision==0.14.0
tqdm==4.64.1
transformers==4.24.0
typing_extensions==4.4.0
urllib3==1.26.12

# package location
--find-links https://download.pytorch.org/whl/torch_stable.html
